+++ Effects/NeckBreak.cs
@@ -1,4 +1,6 @@
 ﻿using GameNetcodeStuff;
+using System;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -12,6 +14,13 @@ namespace MysteryDice.Effects
         public string Tooltip => "Your neck is broken!";
         
         public static int IsNeckBroken = 0;
+
+        public static bool useTimer = true;
+        public static int setTimeMin = 30;
+        public static int setTimeMax = 60;
+        public static float breakTime = 30f;
+        public static bool isTimerRunning = false;
+        
         public void Use()
         {
             Networker.Instance.NeckBreakRandomPlayerServerRpc(GameNetworkManager.Instance.localPlayerController.playerClientId);
@@ -19,6 +28,7 @@ namespace MysteryDice.Effects
 
         public static void BreakNeck()
         {
+            breakTime = UnityEngine.Random.Range(setTimeMin, setTimeMax);
             IsNeckBroken += 1;
         }
         public static void FixNeck()
@@ -27,5 +37,14 @@ namespace MysteryDice.Effects
             Transform cam = GameNetworkManager.Instance.localPlayerController.gameplayCamera.transform;
             cam.eulerAngles = new Vector3(cam.eulerAngles.x, cam.eulerAngles.y, 0f);
         }
+        public static IEnumerator WaitTime()
+        {
+            isTimerRunning = true;
+            MysteryDice.CustomLogger.LogInfo($"Breaking neck for {NeckBreak.breakTime} seconds");
+            yield return new WaitForSeconds(NeckBreak.breakTime);
+            NeckBreak.FixNeck();
+            isTimerRunning = false;
+        }
+
     }
 }