+++ Misc.cs
@@ -172,7 +172,6 @@ namespace MysteryDice
         public static void AdjustWeight(ulong userID, float factor)
         {
             PlayerControllerB player = null;
-
             foreach (GameObject playerPrefab in StartOfRound.Instance.allPlayerObjects)
             {
                 PlayerControllerB playerComp = playerPrefab.GetComponent<PlayerControllerB>();
@@ -182,27 +181,30 @@ namespace MysteryDice
                     break;
                 }
             }
-
             if (player == null)
             {
                 Debug.LogError("Player not found.");
                 return;
             }
 
+            
+            float actualWeight = 0;
             float totalWeight = 0;
 
             foreach (var item in player.ItemSlots)
             {
                 if (item == null || item.itemProperties.weight == 0) continue;
+                actualWeight += item.itemProperties.weight;
                 float displayedWeight = Mathf.RoundToInt(Mathf.Clamp(item.itemProperties.weight - 1f, 0.0f, 100f) * 105f);
                 displayedWeight *= factor;
                 float carryWeight = (displayedWeight / 105f) + 1f;
-                carryWeight = Mathf.Clamp(carryWeight, 1f, 10f); 
+                carryWeight = Mathf.Clamp(carryWeight, 1f, 10f);
                 item.itemProperties.weight = carryWeight;
                 totalWeight += item.itemProperties.weight;
             }
-
-            player.carryWeight = totalWeight;
+            //there might be a better way but I am tired, it is late, and it works so...
+            if (actualWeight != 0) player.carryWeight = totalWeight;
         }
     }
 }
+